requirements.txt
camelot-py[cv]==0.10.1
Scrapy==2.5.0
